fix bug when no sample sheet

+++ pipeline_cron.sh
@@ -40,46 +40,57 @@ function processJobs {
             echo "Not processing run $run"
         else
 
-            # remove spaces from sample sheet
-            sed -i 's/ //g' $raw_write/$instrumentType/$run/SampleSheet.csv
 
-            # modify RTAComplete to prevent cron re-triggering
-            mv $raw_write/$instrumentType/$run/RTAComplete.txt $raw_write/$instrumentType/$run/_RTAComplete.txt
+            if [ -f $raw_write/$instrumentType/$run/SampleSheet.csv ]; then
+
+                # remove spaces from sample sheet
+                sed -i 's/ //g' $raw_write/$instrumentType/$run/SampleSheet.csv
+                
+                 # modify RTAComplete to prevent cron re-triggering
+                 mv $raw_write/$instrumentType/$run/RTAComplete.txt $raw_write/$instrumentType/$run/_RTAComplete.txt
         
-            #sleep 5m 
-            # counting instances of Dragen in SampleSheet
-            set +e
-            is_dragen=$(cat "$path"/SampleSheet.csv | grep "Dragen" | wc -l)
-            set -e
+                 #sleep 5m 
+                 # counting instances of Dragen in SampleSheet
+                 set +e
+                 is_dragen=$(cat "$path"/SampleSheet.csv | grep "Dragen" | wc -l)
+                 set -e
 
-            if [ $is_dragen -gt 0 ]; then
+                 if [ $is_dragen -gt 0 ]; then
 
-                echo "Keyword Dragen found in SampleSheet so executing DragenQC"
-                ssh ch1 "mkdir $fastq_write/$run && cd $fastq_write/$run && sbatch --export=sourceDir=$path /data/diagnostics/pipelines/DragenQC/DragenQC-master/DragenQC.sh"
-            else
+                     echo "Keyword Dragen found in SampleSheet so executing DragenQC"
+                     ssh ch1 "mkdir $fastq_write/$run && cd $fastq_write/$run && sbatch --export=sourceDir=$path /data/diagnostics/pipelines/DragenQC/DragenQC-master/DragenQC.sh"
+                 else
 
-                # launch IlluminaQC for demultiplexing and QC
-                ssh ch1 "mkdir $fastq_write/$run && cd $fastq_write/$run && sbatch -J IlluminaQC-"$run" --export=sourceDir=$path /data/diagnostics/pipelines/IlluminaQC/IlluminaQC-$version/1_IlluminaQC.sh"
-            fi
+                    # launch IlluminaQC for demultiplexing and QC
+                    ssh ch1 "mkdir $fastq_write/$run && cd $fastq_write/$run && sbatch -J IlluminaQC-"$run" --export=sourceDir=$path /data/diagnostics/pipelines/IlluminaQC/IlluminaQC-$version/1_IlluminaQC.sh"
+                 fi
 
-            # check we havent already copied the directory
-            if [ -d "$arc_write/$instrumentType/$run" ]; then
+                 # check we havent already copied the directory
+                 if [ -d "$arc_write/$instrumentType/$run" ]; then
 
-                echo "$arc_write/$instrumentType/$run already exists"
-            else
+                     echo "$arc_write/$instrumentType/$run already exists"
+                 else
 
-                # move run to archive
-                cp -r "$raw_write/$instrumentType/$run" "$arc_write/$instrumentType/$run"
+                     # move run to archive
+                     cp -r "$raw_write/$instrumentType/$run" "$arc_write/$instrumentType/$run"
 
-                touch "$arc_write/$instrumentType/$run"/run_copy_complete.txt
+                     touch "$arc_write/$instrumentType/$run"/run_copy_complete.txt
 
-                # change access permissions
-                chmod -R 755 "$arc_write"/"$instrumentType"/"$run"
-                chmod 777 "$arc_write"/"$instrumentType"/"$run"/SampleSheet.csv
+                     # change access permissions
+                     chmod -R 755 "$arc_write"/"$instrumentType"/"$run"
+                     chmod 777 "$arc_write"/"$instrumentType"/"$run"/SampleSheet.csv
             
-            fi
+                 fi
+
+             else
+
+                 echo "Not running as no sample sheet"
+
+             fi
 
         fi
+
+	
 
     done
 }
